Rename header files to have hpp extension.
