--- include/Stream.h
@@ -1,15 +1,0 @@
-#ifndef _STREAM_H_
-#define _STREAM_H_
-
-#include "StreamTypes.h"
-
-#include "stream/TwoTypeStream.h"
-#include "stream/SplitStream.h"
-#include "stream/Sink.h"
-#include "stream/FilterStream.h"
-#include "stream/MapStream.h"
-#include "stream/Window.h"
-#include "stream/WindowAggregate.h"
-#include "stream/WindowBatch.h"
-
-#endif
