+++ include/Stream.hpp
@@ -1,0 +1,15 @@
+#ifndef _STREAM_H_
+#define _STREAM_H_
+
+#include "StreamTypes.hpp"
+
+#include "stream/TwoTypeStream.hpp"
+#include "stream/SplitStream.hpp"
+#include "stream/Sink.hpp"
+#include "stream/FilterStream.hpp"
+#include "stream/MapStream.hpp"
+#include "stream/Window.hpp"
+#include "stream/WindowAggregate.hpp"
+#include "stream/WindowBatch.hpp"
+
+#endif
