+++ include/StreamTypes.hpp
