+++ include/stream/SplitStream.hpp
@@ -1,7 +1,7 @@
 #ifndef _SPLIT_STREAM_H_
 #define _SPLIT_STREAM_H_
 
-#include "StreamTypes.h"
+#include "StreamTypes.hpp"
 
 template <typename T>
 class SplitStream: public Stream<T> {
