+++ include/stream/Subscriber.hpp
