+++ include/stream/WindowAggregate.hpp
@@ -1,7 +1,7 @@
 #ifndef _WINDOW_AGGREGATE_H_
 #define _WINDOW_AGGREGATE_H_
 
-#include <StreamTypes.h>
+#include <StreamTypes.hpp>
 #include <list>
 
 template <typename INPUT, typename OUTPUT>
